refactor(*): ファイルを分割して管理
ファイルを分割して管理し、esbuildでバンドルした後にpushするように設定
